Added invoice generation and signature upload

+++ AgrifoodManagement.Web/Models/Shop/ProductViewModel.cs
@@ -16,7 +16,7 @@ namespace AgrifoodManagement.Web.Models.Shop
         public MeasurementUnit? UnitOfMeasurement { get; set; }
         public string UnitOfMeasurementText => UnitOfMeasurement.ToString();
         public DateTime ExpiryDate { get; set; }
-        public decimal Rating { get; set; }
+        public double Rating { get; set; }
         public int ReviewCount { get; set; }
         public bool IsFavorited { get; set; }
         public bool IsExpiringSoon => ExpiryDate.Subtract(DateTime.Now).TotalDays <= 3;
